services/news-analyzer/requirements.txt
# services/news-analyzer/requirements.txt
redis>=5.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
mysql-connector-python>=8.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0

